refactor: extract shared formatNumberValue into tokenValueUtils
